pyproject.toml
[project]
name = "director-agent"
version = "0.1.0"
description = "An autonomous movie studio agent orchestrating Gemini, Veo, and TTS tools."
authors = [{name = "Chuck Forsyth", email = "forsythc@ucr.edu"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "typer>=0.9.0",
    "rich>=13.0",
    "ffmpeg-python>=0.2.0",
    "google-generativeai>=0.3.0",
]

[project.scripts]
director = "director_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
